Refactor into app package with routers
