+++ tests/test_image_cache.py
@@ -5,7 +5,7 @@ from cachetools import TTLCache
 import httpx
 
 
-import main
+import app.routes.cards as cards_routes
 
 
 def test_image_url_cache(monkeypatch):
@@ -21,11 +21,11 @@ def test_image_url_cache(monkeypatch):
 
         return R()
 
-    monkeypatch.setattr(main, "_image_cache", TTLCache(maxsize=10, ttl=10))
+    monkeypatch.setattr(cards_routes, "_image_cache", TTLCache(maxsize=10, ttl=10))
     monkeypatch.setattr(httpx.AsyncClient, "head", dummy_head)
 
-    url1 = asyncio.run(main._image_url("de", "A2a", "001"))
-    url2 = asyncio.run(main._image_url("de", "A2a", "001"))
+    url1 = asyncio.run(cards_routes._image_url("de", "A2a", "001"))
+    url2 = asyncio.run(cards_routes._image_url("de", "A2a", "001"))
     assert url1 == url2
     assert len(calls) == 1
 
@@ -37,11 +37,11 @@ def test_image_url_timeout(monkeypatch, caplog):
         raise httpx.TimeoutException("boom")
 
     cache = TTLCache(maxsize=10, ttl=10)
-    monkeypatch.setattr(main, "_image_cache", cache)
+    monkeypatch.setattr(cards_routes, "_image_cache", cache)
     monkeypatch.setattr(httpx.AsyncClient, "head", timeout_head)
 
     caplog.set_level(logging.ERROR)
-    url = asyncio.run(main._image_url("de", "A2a", "001"))
+    url = asyncio.run(cards_routes._image_url("de", "A2a", "001"))
     assert url.endswith("low.webp")
     high = "https://assets.tcgdex.net/de/tcgp/A2a/001/high.webp"
     assert cache[high] is False
